feat : 마감된 투표 페이지 mocking api 연결
