+++ src/components/common/Router.tsx
@@ -18,7 +18,6 @@ const Router = () => (
         <Route path="/library" element={<Library />} />
         <Route path="/login" element={<Login />} />
         <Route path="/signup" element={<Signup />} />
-        {/* <Route path="/playerlanding" element={<PlayerLanding />} /> */}
         <Route path="/vote/:voteId" element={<PlayerLanding />} />
         <Route path="/makerlanding" element={<MakerLanding />} />
         <Route path="/voting" element={<Voting />} />
