+++ src/lib/api/playerLanding.ts
@@ -1,5 +1,7 @@
 import axios from 'axios';
 
+import { VoteData } from '../../types/vote';
+
 export const client = axios.create({
   baseURL: 'https://ed76eca9-2182-4ba8-9458-3321e7958ab4.mock.pstmn.io',
   headers: {
@@ -7,29 +9,12 @@ export const client = axios.create({
   },
 });
 
-export interface VoteInfo {
-  // username: string;
-  vote_id: number;
-  vote_status: boolean;
-  vote_title: string;
-  Picture: PictureData[];
-}
-
-export interface PictureData {
-  picture_id: number;
-  url: string;
-}
-
-export interface VoteData {
-  status: number;
-  data: VoteInfo;
-}
-
 export const getVoteStatus = async (voteId: number) => {
   try {
     const { data } = await client.get<VoteData>(`/vote/${voteId}`);
     console.log(data.status);
     if (data.status === 200) {
+      console.log(data);
       return data.data;
     }
     if (data.status === 400) {
